+++ crates/cli/src/error.rs
@@ -50,7 +50,7 @@ pub enum ProtoCliError {
     #[error("Upgrading proto requires an internet connection!")]
     UpgradeRequiresInternet,
 
-    #[diagnostic(code(proto::cli::proto_running))]
+    #[diagnostic(code(proto::cli::running_process))]
     #[error("Unable to upgrade as an instance of proto is currently running with the process ID {}. Please stop this process then try again.", .pid.to_string().style(Style::Symbol))]
     CannotUpgradeProtoRunning { pid: u32 },